default values removed
